requirements.txt
numpy==1.21.3
pandas==1.3.4
mip==1.13.0
pyyaml==6.0
dwave-ocean-sdk==4.2.0
openjij==0.4.5
matplotlib==3.4.3
